+++ src/prebuilt/ver1_46/units.rs
@@ -7,9 +7,9 @@ pub struct Units {}
 impl Units {
     pub fn template() -> Token {
         let mut root = LinkedHashMap::with_capacity(5);
-        root.push_back("number_of_unit_sections", 9 as u32);
+        root.push_back("number_of_unit_sections", 9_u32);
         root.push_back("player_data_3", vec![PlayerDataThree::template(); 8]);
-        root.push_back("number_of_players", 9 as u32);
+        root.push_back("number_of_players", 9_u32);
         root.push_back("player_data_4", vec![PlayerDataFour::template(); 8]);
         root.push_back("players_units", vec![PlayerUnits::template(); 1]);
         root.patchs.insert(
